requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
pycryptodome==3.19.0